+++ tests/test_print.py
@@ -19,27 +19,27 @@ class TestPrint(unittest.TestCase):
             " 'decimal_separator': '.',\n"
             " 'lower_separator': '',\n"
             " 'sign_mode': '-',\n"
-            " 'fill_mode': ' ',\n"
-            " 'top_dig_place': 0,\n"
+            " 'fill_char': ' ',\n"
+            " 'left_pad_dec_place': 0,\n"
             " 'exp_format': 'standard',\n"
             " 'extra_si_prefixes': {},\n"
             " 'extra_iec_prefixes': {},\n"
             " 'extra_parts_per_forms': {},\n"
             " 'capitalize': False,\n"
-            " 'superscript_exp': False,\n"
+            " 'superscript': False,\n"
             " 'latex': False,\n"
             " 'nan_inf_exp': False,\n"
-            " 'bracket_unc': False,\n"
+            " 'paren_uncertainty': False,\n"
             " 'pdg_sig_figs': False,\n"
-            " 'val_unc_match_widths': False,\n"
-            " 'bracket_unc_remove_seps': False,\n"
-            " 'unc_pm_whitespace': True}\n"
+            " 'left_pad_matching': False,\n"
+            " 'paren_uncertainty_separators': True,\n"
+            " 'pm_whitespace': True}\n"
         )
         self.assertEqual(actual_printout, expected_printout)
 
     def test_unrendered_options_repr(self):
         with redirect_stdout(StringIO()) as sout:  # noqa: SIM117
-            with GlobalDefaultsContext(top_dig_place=3, capitalize=True):
+            with GlobalDefaultsContext(left_pad_dec_place=3, capitalize=True):
                 print_global_defaults()
 
         actual_printout = sout.getvalue()
@@ -52,20 +52,20 @@ class TestPrint(unittest.TestCase):
             " 'decimal_separator': '.',\n"
             " 'lower_separator': '',\n"
             " 'sign_mode': '-',\n"
-            " 'fill_mode': ' ',\n"
-            " 'top_dig_place': 3,\n"
+            " 'fill_char': ' ',\n"
+            " 'left_pad_dec_place': 3,\n"
             " 'exp_format': 'standard',\n"
             " 'extra_si_prefixes': {},\n"
             " 'extra_iec_prefixes': {},\n"
             " 'extra_parts_per_forms': {},\n"
             " 'capitalize': True,\n"
-            " 'superscript_exp': False,\n"
+            " 'superscript': False,\n"
             " 'latex': False,\n"
             " 'nan_inf_exp': False,\n"
-            " 'bracket_unc': False,\n"
+            " 'paren_uncertainty': False,\n"
             " 'pdg_sig_figs': False,\n"
-            " 'val_unc_match_widths': False,\n"
-            " 'bracket_unc_remove_seps': False,\n"
-            " 'unc_pm_whitespace': True}\n"
+            " 'left_pad_matching': False,\n"
+            " 'paren_uncertainty_separators': True,\n"
+            " 'pm_whitespace': True}\n"
         )
         self.assertEqual(actual_printout, expected_printout)