+++ tests/test_val_unc_formatter.py
@@ -18,30 +18,41 @@ class TestFormatting(unittest.TestCase):
                 ):
                     self.assertEqual(snum_str, expected_val_unc_str)
 
-    def test_bracket_unc(self):
+    def test_paren_uncertainty(self):
         cases_list = [
             (
                 (123.456, 0.789),
                 [
-                    (Formatter(bracket_unc=True), "123.456(789)"),
+                    (Formatter(paren_uncertainty=True), "123.456(789)"),
                     (
-                        Formatter(exp_mode="scientific", bracket_unc=True),
+                        Formatter(exp_mode="scientific", paren_uncertainty=True),
                         "(1.23456(789))e+02",
                     ),
                     (
-                        Formatter(exp_mode="engineering", bracket_unc=True),
+                        Formatter(exp_mode="engineering", paren_uncertainty=True),
                         "(123.456(789))e+00",
                     ),
                     (
-                        Formatter(exp_mode="engineering_shifted", bracket_unc=True),
+                        Formatter(
+                            exp_mode="engineering_shifted",
+                            paren_uncertainty=True,
+                        ),
                         "(0.123456(789))e+03",
                     ),
                     (
-                        Formatter(exp_mode="scientific", exp_val=+1, bracket_unc=True),
+                        Formatter(
+                            exp_mode="scientific",
+                            exp_val=+1,
+                            paren_uncertainty=True,
+                        ),
                         "(12.3456(789))e+01",
                     ),
                     (
-                        Formatter(exp_mode="scientific", exp_val=-1, bracket_unc=True),
+                        Formatter(
+                            exp_mode="scientific",
+                            exp_val=-1,
+                            paren_uncertainty=True,
+                        ),
                         "(1234.56(7.89))e-01",
                     ),
                 ],
@@ -50,36 +61,36 @@ class TestFormatting(unittest.TestCase):
 
         self.run_val_unc_formatter_cases(cases_list)
 
-    def test_bracket_unc_invalid_unc(self):
+    def test_paren_unc_invalid_unc(self):
         cases_list = [
             (
                 (123, 0),
                 [
-                    (Formatter(bracket_unc=True), "123(0)"),
+                    (Formatter(paren_uncertainty=True), "123(0)"),
                 ],
             ),
             (
                 (-123, 0),
                 [
-                    (Formatter(bracket_unc=True), "-123(0)"),
+                    (Formatter(paren_uncertainty=True), "-123(0)"),
                 ],
             ),
             (
                 (123, float("nan")),
                 [
-                    (Formatter(bracket_unc=True), "123(nan)"),
+                    (Formatter(paren_uncertainty=True), "123(nan)"),
                 ],
             ),
             (
                 (123, float("inf")),
                 [
-                    (Formatter(bracket_unc=True), "123(inf)"),
+                    (Formatter(paren_uncertainty=True), "123(inf)"),
                 ],
             ),
             (
                 (0, 0),
                 [
-                    (Formatter(bracket_unc=True), "0(0)"),
+                    (Formatter(paren_uncertainty=True), "0(0)"),
                 ],
             ),
         ]
@@ -93,7 +104,7 @@ class TestFormatting(unittest.TestCase):
                 [
                     (
                         Formatter(
-                            bracket_unc=True,
+                            paren_uncertainty=True,
                             exp_mode="engineering",
                             exp_format="prefix",
                         ),
@@ -101,7 +112,7 @@ class TestFormatting(unittest.TestCase):
                     ),
                     (
                         Formatter(
-                            bracket_unc=True,
+                            paren_uncertainty=True,
                             exp_mode="engineering_shifted",
                             exp_format="prefix",
                         ),
@@ -114,7 +125,7 @@ class TestFormatting(unittest.TestCase):
                 [
                     (
                         Formatter(
-                            bracket_unc=True,
+                            paren_uncertainty=True,
                             exp_mode="fixed_point",
                             exp_format="prefix",
                         ),
@@ -122,7 +133,7 @@ class TestFormatting(unittest.TestCase):
                     ),
                     (
                         Formatter(
-                            bracket_unc=True,
+                            paren_uncertainty=True,
                             exp_mode="engineering",
                             exp_format="prefix",
                         ),
@@ -130,7 +141,7 @@ class TestFormatting(unittest.TestCase):
                     ),
                     (
                         Formatter(
-                            bracket_unc=True,
+                            paren_uncertainty=True,
                             exp_mode="engineering_shifted",
                             exp_format="prefix",
                         ),
@@ -143,7 +154,7 @@ class TestFormatting(unittest.TestCase):
                 [
                     (
                         Formatter(
-                            bracket_unc=True,
+                            paren_uncertainty=True,
                             exp_mode="engineering",
                             exp_format="parts_per",
                         ),
@@ -151,7 +162,7 @@ class TestFormatting(unittest.TestCase):
                     ),
                     (
                         Formatter(
-                            bracket_unc=True,
+                            paren_uncertainty=True,
                             exp_mode="engineering",
                             exp_format="prefix",
                         ),
@@ -159,7 +170,7 @@ class TestFormatting(unittest.TestCase):
                     ),
                     (
                         Formatter(
-                            bracket_unc=True,
+                            paren_uncertainty=True,
                             exp_mode="engineering_shifted",
                             exp_format="parts_per",
                             add_ppth_form=True,
@@ -184,7 +195,7 @@ class TestFormatting(unittest.TestCase):
                     (
                         Formatter(
                             exp_mode="percent",
-                            bracket_unc=True,
+                            paren_uncertainty=True,
                             lower_separator="_",
                         ),
                         "(12.345_678(255))%",
@@ -195,21 +206,25 @@ class TestFormatting(unittest.TestCase):
 
         self.run_val_unc_formatter_cases(cases_list)
 
-    def test_bracket_unc_remove_dec_symb(self):
+    def test_paren_unc_separators(self):
         cases_list = [
             (
                 (123.456, 0.789),
                 [
                     (
-                        Formatter(exp_mode="scientific", exp_val=-1, bracket_unc=True),
+                        Formatter(
+                            exp_mode="scientific",
+                            exp_val=-1,
+                            paren_uncertainty=True,
+                        ),
                         "(1234.56(7.89))e-01",
                     ),
                     (
                         Formatter(
                             exp_mode="scientific",
                             exp_val=-1,
-                            bracket_unc_remove_seps=True,
-                            bracket_unc=True,
+                            paren_uncertainty_separators=False,
+                            paren_uncertainty=True,
                         ),
                         "(1234.56(789))e-01",
                     ),
@@ -219,7 +234,11 @@ class TestFormatting(unittest.TestCase):
                 (0.789, 123.456),
                 [
                     (
-                        Formatter(exp_mode="scientific", exp_val=-1, bracket_unc=True),
+                        Formatter(
+                            exp_mode="scientific",
+                            exp_val=-1,
+                            paren_uncertainty=True,
+                        ),
                         "(7.89(1234.56))e-01",
                     ),
                     # Don't remove "embedded" decimal unless val > unc.
@@ -227,10 +246,67 @@ class TestFormatting(unittest.TestCase):
                         Formatter(
                             exp_mode="scientific",
                             exp_val=-1,
-                            bracket_unc_remove_seps=True,
-                            bracket_unc=True,
+                            paren_uncertainty_separators=False,
+                            paren_uncertainty=True,
                         ),
                         "(7.89(1234.56))e-01",
+                    ),
+                ],
+            ),
+            (
+                (1.2, float("nan")),
+                [
+                    (
+                        Formatter(
+                            paren_uncertainty=True,
+                            paren_uncertainty_separators=False,
+                        ),
+                        "1.2(nan)",
+                    ),
+                    (
+                        Formatter(
+                            paren_uncertainty=True,
+                            paren_uncertainty_separators=True,
+                        ),
+                        "1.2(nan)",
+                    ),
+                ],
+            ),
+            (
+                (float("nan"), 1.2),
+                [
+                    (
+                        Formatter(
+                            paren_uncertainty=True,
+                            paren_uncertainty_separators=False,
+                        ),
+                        "nan(1.2)",
+                    ),
+                    (
+                        Formatter(
+                            paren_uncertainty=True,
+                            paren_uncertainty_separators=True,
+                        ),
+                        "nan(1.2)",
+                    ),
+                ],
+            ),
+            (
+                (float("nan"), float("nan")),
+                [
+                    (
+                        Formatter(
+                            paren_uncertainty=True,
+                            paren_uncertainty_separators=True,
+                        ),
+                        "nan(nan)",
+                    ),
+                    (
+                        Formatter(
+                            paren_uncertainty=True,
+                            paren_uncertainty_separators=False,
+                        ),
+                        "nan(nan)",
                     ),
                 ],
             ),
@@ -238,26 +314,26 @@ class TestFormatting(unittest.TestCase):
 
         self.run_val_unc_formatter_cases(cases_list)
 
-    def test_unc_pm_whitespace(self):
+    def test_pm_whitespace(self):
         cases_list = [
             (
                 (123.456, 0.789),
                 [
-                    (Formatter(unc_pm_whitespace=True), "123.456 ± 0.789"),
-                    (Formatter(unc_pm_whitespace=False), "123.456±0.789"),
+                    (Formatter(pm_whitespace=True), "123.456 ± 0.789"),
+                    (Formatter(pm_whitespace=False), "123.456±0.789"),
                 ],
             ),
         ]
 
         self.run_val_unc_formatter_cases(cases_list)
 
-    def test_superscript_exp(self):
+    def test_superscript(self):
         cases_list = [
             (
                 (789, 0.01),
                 [
                     (
-                        Formatter(exp_mode="scientific", superscript_exp=True),
+                        Formatter(exp_mode="scientific", superscript=True),
                         "(7.8900 ± 0.0001)×10²",
                     ),
                 ],
@@ -302,7 +378,7 @@ class TestFormatting(unittest.TestCase):
                         Formatter(
                             lower_separator="_",
                             exp_mode="percent",
-                            bracket_unc=True,
+                            paren_uncertainty=True,
                             latex=True,
                         ),
                         r"\left(12.345\_678\left(255\right)\right)\%",