+++ CHANGELOG.rst
@@ -30,12 +30,40 @@ Added
 Changed
 ^^^^^^^
 
+* **[BREAKING]** Renamed multiple options.
+
+    * ``top_dig_place`` renamed to ``left_pad_dec_place``.
+    * ``superscript_exp`` renamed to ``superscript``.
+    * ``bracket_unc`` renamed to ``paren_uncertainty``.
+    * ``bracket_unc_remove_seps`` renamed to
+      ``paren_uncertainty_separators``. This change is associated with a
+      a reversal of the Boolean logic on the option.
+    * ``val_unc_match_widths`` renamed to ``left_pad_matching``.
+    * ``unc_pm_whitespace`` renamed to ``pm_whitespace``.
+
+* **[BREAKING]** Previously specifying any left pad decimal place using
+  the ``sciform`` FSML resulted in setting ``left_pad_matching=True`` so
+  that ``print(f"{SciNum(123.456, 0.789):0}")`` resulted in
+  ``"123.456 ± 000.789"``.
+  Now the FSML has no impact on ``left_pad_matching``.
+  Now, similar to many other options, the global setting for
+  ``left_pad_matching`` will always be used when formatting using the
+  FSML.
+  Under the default global options (``left_pad_matching=False``)
+  ``print(f"{SciNum(123.456, 0.789):0}")`` results in
+  ``"123.456 ± 0.789"``.
 * Implemented `ruff <https://docs.astral.sh/ruff/>`_ linting and
   formatting in codebase and integration automation.
 * Refactored code for adding separators.
 * Refactored formatting and formatting utilities to simplify functions
   and make the algorithm easier to follow.
-* More aggressively filter JetBrains .idea folder from version control.
+* More aggressively filter JetBrains ``.idea/`` folder from version control.
+
+Fixed
+^^^^^
+
+* Fixed a bug involving removing separators in parentheses uncertainty
+  mode when at least one of the value and uncertainty were non-finite.
 
 ----